rockchip: add PINE64 SOQuartz Model A board
The PINE64 SOQuartz Model A board is a carrier board for the
SOQuartz CM4-compatible compute module. It exposes PCIe,
ethernet, USB, HDMI, CSI, DSI, eDP and a 40 pin GPIO header,
and is powered by 12V DC.

The board config enabled FDT overlays as this is a development
board people would reasonably want to plug things into. As this
bloats the FDT size to beyond a magic ~62000 byte limit, the
vendor ATF binary chokes on it, so SPL_ATF_NO_PLATFORM_PARAM
needs to be set for it to boot.

Signed-off-by: Nicolas Frattaroli <frattaroli.nicolas@gmail.com>